refactor : step5 리팩토링 패키지 변경
