+++ src/main/java/racingcar/RacingGameApplication.java
@@ -1,6 +1,8 @@
 package racingcar;
 
-import racingcar.view.RacingCarView;
+import racingcar.domain.Car;
+import racingcar.domain.RacingGame;
+import racingcar.view.RacingGameResultView;
 import racingcar.view.RacingGameInputView;
 
 import java.util.List;
@@ -9,10 +11,10 @@ public class RacingGameApplication {
     public static void main(String[] args) {
         String carNames = RacingGameInputView.inputCarNumber();
         int moveNumber = RacingGameInputView.inputMoveNumber();
-        RacingCarView.printMessage();
+        RacingGameResultView.printMessage();
         RacingGame racingGame = new RacingGame(carNames);
         List<Car> cars = racingGame.race(moveNumber);
-        RacingCarView.printRacingCar(cars);
-        RacingCarView.printWinner(racingGame.getWinner());
+        RacingGameResultView.printRacingCar(cars);
+        RacingGameResultView.printWinner(racingGame.getWinner());
     }
 }
